Refactor flags, CLI descriptions (#196)
* initial commit for refactoring flags, CLI descriptions

* fix getAllowanceCmd name

* Refactor CLI commands

* fix tokenId flag name + description

* fix tokenId flag name

* add NumberOfBlocks flag variable

* rename hash-list comamnd + refactor

* fix gasLimit + gasPrice flag formatting

* fix flag formatting

* fix flag formatting

* fix flag formatting

* fix flag formatting

* fix flag formatting

* fix flag/description formatting

* fix required flags order

* fix flag formatting

* update long description

* update short/long description

* update short description

* fix description + flag formatting

* update short description

* update short description

* update short description

* fix short/long description

* update short/long description

* update short description

* update short/long description

* update short/long descriptions

* fix short description

* fix short description

* fix long description

* update short description

* update short description

* fix error capitalization warnings

* fix description

* update long description

* fix short/long descriptions + flag formatting

* update flag descriptions

* add tx hash logging + formatting

* fix hash-list command

* refactor flag names

* refactor flag names

* refactor cli flag name format

* Fix relayer-threshold flag typo

* reformat flags

* refactor flag names

* fix description

* update readme

* fix domain flag + relabel token-contract flag

* fix logger test

* fix flag name

* fix set-burn/withdraw command flags

* fix logger test

* Set all domain ids to uint8

+++ chains/evm/cli/bridge/flagVars.go
@@ -9,7 +9,7 @@ import (
 var (
 	Bridge          string
 	DataHash        string
-	DomainID        uint64
+	DomainID        uint8
 	DepositNonce    uint64
 	Handler         string
 	ResourceID      string
